backend/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
Django==5.0.8
djangorestframework==3.15.2
pillow==10.3.0
platformdirs==4.2.2
PyJWT==2.8.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
pytest-django
<<<<<<< HEAD
django-environ==0.11.2
=======
>>>>>>> upstream/main
